requirements.txt
tensorflow
pillow
numpy
streamlit
opencv-python